aider: Added logging configuration to output debug messages to the console.
# Aider chat conversation:

USER: this is what I see '/Users/jonbennett/Dropbox/Coding/FDAS3/fdas/mainapp/views.py changed, reloading.
Watching for file changes with StatReloader
Performing system checks...

System check identified no issues (0 silenced).
December 18, 2023 - 13:47:29
Django version 5.0, using settings 'fdas.settings'
Starting development server at http://127.0.0.1:8000/
Quit the server with CONTROL-C.'

ASSISTANT: The message you're seeing indicates that the Django development server has detected changes to `views.py` and has reloaded the file. This is expected behavior when you make changes to a file in a Django project while the development server is running.

The server is running without any system checks issues, which is good. However, since you're still not seeing data in the table, we need to check the logs for the debug messages we added to the `get_data` view.

Please look at the server console output after you make a request to `/get_data/`. You should see the debug messages that were added to the `get_data` view. These messages will show the data fetched from the database and the data that is being sent as JSON.

If you don't see any debug messages, make sure that the logging level is set to `DEBUG` in your Django settings. You can do this by adding the following configuration to your `fdas/fdas/settings.py` file:

```python
LOGGING = {
    'version': 1,
    'disable_existing_loggers': False,
    'handlers': {
        'console': {
            'level': 'DEBUG',
            'class': 'logging.StreamHandler',
        },
    },
    'loggers': {
        '': {
            'handlers': ['console'],
            'level': 'DEBUG',
            'propagate': True,
        },
    },
}
```

If you need a diff to add this logging configuration, please let me know. Once the logging level is set correctly, you should be able to see the debug messages and determine whether the data is being fetched and serialized correctly. If the data is correct in the logs but not displaying on the front end, the issue might be in the JavaScript code or the HTML template.

USER: please add that code

ASSISTANT: Here is the diff to add the logging configuration to your `fdas/fdas/settings.py` file:

```diff
--- fdas/fdas/settings.py
+++ fdas/fdas/settings.py
@@ ... @@
 # Default primary key field type
 # https://docs.djangoproject.com/en/5.0/ref/settings/#default-auto-field

 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
+
+LOGGING = {
+    'version': 1,
+    'disable_existing_loggers': False,
+    'handlers': {
+        'console': {
+            'level': 'DEBUG',
+            'class': 'logging.StreamHandler',
+        },
+    },
+    'loggers': {
+        '': {
+            'handlers': ['console'],
+            'level': 'DEBUG',
+            'propagate': True,
+        },
+    },
+}
```

This will configure Django to output debug messages to the console. After applying this change, you should be able to see the debug messages from the `get_data` view in the server console when you make a request to `/get_data/`. This will help you to troubleshoot the issue further.